refactor: optimize org-member DTOs and decorators
