+++ src/admin/org-member/dto/index.ts
@@ -3,6 +3,6 @@ export * from './update-org-member.dto';
 export * from './update-member-status.dto';
 export * from './update-member-departments.dto';
 export * from './batch-import-members.dto';
-export * from './pagination.dto';
+export * from './query-org-member.dto';
 export * from './org-member.dto';
-export * from './batch-import-members.dto';
+export * from './member-role-option.dto';
