+++ src/admin/org-member/dto/org-member.dto.ts
@@ -221,25 +221,6 @@ export class OrgMemberListItemDto {
   primaryDept: OrgMemberListPrimaryDeptInfo | null;
 }
 
-export class MemberRoleOptionDto {
-  @ApiProperty() id: string;
-  @ApiProperty() userId: string;
-  @ApiPropertyOptional({ type: String, nullable: true }) displayName:
-    | string
-    | null;
-  @ApiPropertyOptional({ type: String, nullable: true }) email: string | null;
-  @ApiPropertyOptional({ type: String, nullable: true }) avatar: string | null;
-  @ApiProperty({ type: [String] }) departmentNames: string[];
-  @ApiProperty({ type: [String] }) roleIds: string[];
-}
-
-export class MemberRoleOptionListResponse {
-  @ApiProperty({ type: [MemberRoleOptionDto] }) items: MemberRoleOptionDto[];
-  @ApiProperty() total: number;
-  @ApiProperty() page: number;
-  @ApiProperty() pageSize: number;
-  @ApiProperty() totalPages: number;
-}
 
 export class OrgMemberDepartmentInfo {
   @ApiProperty({
@@ -333,54 +314,3 @@ export class OrgMemberListResponse {
   totalPages: number;
 }
 
-export class BatchImportFailure {
-  @ApiProperty({
-    description: '成员在批量请求中的序号（从 0 开始）',
-    example: 0,
-  })
-  index: number;
-
-  @ApiPropertyOptional({
-    description: '脱敏邮箱',
-    example: 'zh***@example.com',
-  })
-  email?: string;
-
-  @ApiPropertyOptional({
-    description: '脱敏手机号',
-    example: '138****8000',
-  })
-  phone?: string;
-
-  @ApiProperty({
-    description: '稳定错误码',
-    example: 'CONFLICT',
-  })
-  code: string;
-
-  @ApiProperty({
-    description: '失败原因',
-    example: 'User is already a member of this organization',
-  })
-  reason: string;
-}
-
-export class BatchImportResponse {
-  @ApiProperty({
-    description: '成功导入的成员数量',
-    example: 8,
-  })
-  successCount: number;
-
-  @ApiProperty({
-    description: '失败的成员数量',
-    example: 2,
-  })
-  failureCount: number;
-
-  @ApiProperty({
-    description: '失败的成员列表',
-    type: [BatchImportFailure],
-  })
-  failures: BatchImportFailure[];
-}
